+++ docs/docs/.vitepress/utils/sidebar.ts
@@ -32,7 +32,6 @@ export const sidebar = {
       link: '/components/',
       items: [
         { text: 'Button 按钮', link: '/components/button' },
-        { text: 'Icon 图标', link: '/components/icon' },
         { text: 'Svg Icon 图标', link: '/components/svg-icon' },
         { text: 'Link 链接', link: '/components/link' },
         { text: 'Text 文本', link: '/components/text' },
@@ -86,7 +85,8 @@ export const sidebar = {
         { text: 'Dialog 对话框', link: '/components/dialog' },
         { text: 'Drawer 抽屉', link: '/components/drawer' },
         { text: 'Tooltip 消息提示', link: '/components/tooltip' },
-        { text: 'Popup 弹出窗口', link: '/components/popup' }
+        { text: 'Popup 弹出窗口', link: '/components/popup' },
+        { text: 'Swap切换', link: '/components/swap' }
       ]
     },
     {