+++ packages/fighting-design/__test__/avatar.spec.ts
@@ -62,26 +62,26 @@ describe('FAvatar', () => {
     expect(wrapper.attributes('style')).toContain('red')
   })
 
-  test('icon', () => {
-    const wrapper = mount(FAvatar, {
-      props: { icon: 'f-icon-default-template' }
-    })
-    expect(wrapper.find('.f-icon').classes()).toContain(
-      'f-icon-default-template'
-    )
-  })
+  // test('icon', () => {
+  //   const wrapper = mount(FAvatar, {
+  //     props: { icon: 'f-icon-default-template' }
+  //   })
+  //   expect(wrapper.find('.f-icon').classes()).toContain(
+  //     'f-icon-default-template'
+  //   )
+  // })
 
-  test('icon-size', () => {
-    const wrapper = mount(FAvatar, {
-      props: { icon: 'f-icon-default-template', fontSize: '12px' }
-    })
-    expect(wrapper.find('.f-icon').attributes('style')).toContain('12px')
-  })
+  // test('icon-size', () => {
+  //   const wrapper = mount(FAvatar, {
+  //     props: { icon: 'f-icon-default-template', fontSize: '12px' }
+  //   })
+  //   expect(wrapper.find('.f-icon').attributes('style')).toContain('12px')
+  // })
 
-  test('icon-color', () => {
-    const wrapper = mount(FAvatar, {
-      props: { icon: 'f-icon-default-template', fontColor: 'red' }
-    })
-    expect(wrapper.find('.f-icon').attributes('style')).toContain('red')
-  })
+  // test('icon-color', () => {
+  //   const wrapper = mount(FAvatar, {
+  //     props: { icon: 'f-icon-default-template', fontColor: 'red' }
+  //   })
+  //   expect(wrapper.find('.f-icon').attributes('style')).toContain('red')
+  // })
 })