+++ packages/fighting-design/breadcrumb-item/src/breadcrumb-item.vue
@@ -1,34 +1,24 @@
 <script lang="ts" setup name="FBreadcrumbItem">
-  import { inject, toRefs } from 'vue'
+  import { inject, toRefs, h } from 'vue'
   import { BreadcrumbPropsKey } from '../../breadcrumb/src/breadcrumb'
-  import { FIcon } from '../../icon'
+  import { FSvgIcon } from '../../svg-icon'
+  import FIconChevronRight from '../../_components/svg/f-icon-chevron-right.vue'
   import type { BreadcrumbPropsType } from '../../breadcrumb/src/breadcrumb'
 
   const breadcrumbContext = inject(BreadcrumbPropsKey) as BreadcrumbPropsType
-  const { separator, separatorIcon, itemColor, separatorColor } =
-    toRefs(breadcrumbContext)
+  const { separator, itemColor, separatorColor } = toRefs(breadcrumbContext)
 </script>
 
 <template>
   <span class="f-breadcrumb-item">
-    <span class="f-breadcrumb-item-inner" :style="{ color: itemColor }">
+    <span class="f-breadcrumb-item__inner" :style="{ color: itemColor }">
       <slot />
     </span>
 
-    <f-icon
-      v-if="separatorIcon"
-      class="f-breadcrumb-item-separator"
-      :icon="separatorIcon"
+    <f-svg-icon
+      :icon="separator || h(FIconChevronRight)"
       :color="separatorColor"
+      :size="16"
     />
-
-    <span
-      v-else
-      class="f-breadcrumb-item-separator"
-      role="presentation"
-      :style="{ color: separatorColor }"
-    >
-      {{ separator }}
-    </span>
   </span>
 </template>