+++ packages/fighting-design/notification/src/notification.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup name="FMessage">
   import { computed, onMounted, ref, isVNode, nextTick } from 'vue'
-  import { FIcon } from '../../icon'
+  import { FCloseBtn } from '../../close-btn'
   import { isString } from '../../_utils'
   import { Props, Emits } from './notification'
   import { notificationDefaultIcon } from '../../_model/notification/type'
@@ -129,8 +129,8 @@
 
 <template>
   <transition
-    :name="`f-notification-fade` + (isRight ? '-right' : '-left')"
     mode="out-in"
+    :name="`f-notification-fade` + (isRight ? '-right' : '-left')"
     @before-leave="closeMessageEnd"
     @after-leave="$emit('destroy')"
   >
@@ -144,9 +144,10 @@
     >
       <!-- icon -->
       <div v-if="showIcon && _icon" class="f-notification--icon">
-        <component :is="_icon" v-if="isVNode(_icon)" />
-        <f-icon v-if="isString(_icon)" size="24px" :icon="(_icon as string)" />
+        <component :is="_icon" v-if="isVNode(_icon)" :size="28" />
+        <!-- <f-icon v-if="isString(_icon)" size="24px" :icon="(_icon as string)" /> -->
       </div>
+
       <!-- 主体内容 -->
       <div class="f-notification--info">
         <!-- 标题 -->
@@ -156,14 +157,16 @@
             {{ title }}
           </h3>
         </div>
+
         <!-- 消息文本 -->
         <component :is="message" v-if="isVNode(message)" />
         <div v-else class="f-notification--text">
           {{ message }}
         </div>
       </div>
+
       <!-- 关闭按钮 -->
-      <div
+      <!-- <div
         v-if="prop.close"
         class="f-notification--close"
         @click="closeMessage"
@@ -171,6 +174,17 @@
         <component :is="closeBtn" v-if="isVNode(closeBtn)" />
         <span v-else-if="closeBtn && isString(closeBtn)">{{ closeBtn }}</span>
         <f-icon v-else size="16px" icon="f-icon-close" />
+      </div> -->
+      <!-- 关闭按钮 -->
+      <div
+        v-if="prop.close"
+        class="f-notification--close"
+        @click="closeMessage"
+      >
+        <template v-if="isString(closeBtn)">{{ closeBtn }}</template>
+        <f-close-btn v-else :size="16">
+          <!-- <component :is="closeBtn" /> -->
+        </f-close-btn>
       </div>
     </div>
   </transition>