+++ packages/fighting-design/swap/src/interface.d.ts
@@ -1,0 +1,5 @@
+export type { SwapPropsType } from './swap'
+
+export type SwapSizeType = 'large' | 'middle' | 'small'
+
+export type SwapType = 'sound' | 'theme' | 'rotate' | 'flip' | 'favorites'