+++ packages/fighting-design/switch/src/switch.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup name="FSwitch">
   import { computed } from 'vue'
   import { Props, Emits } from './switch'
-  import { FIcon } from '../../icon'
+  import { FSvgIcon } from '../../svg-icon'
   import type {
     OrdinaryFunctionInterface as a,
     ClassListInterface as b
@@ -11,10 +11,8 @@
   const prop = defineProps(Props)
   const emit = defineEmits(Emits)
 
-  const changeSwitch: a = (): void => {
-    if (prop.disabled) {
-      return
-    }
+  const handleClick: a = (): void => {
+    if (prop.disabled) return
     emit('update:modelValue', !prop.modelValue)
     emit('change', !prop.modelValue)
   }
@@ -60,10 +58,10 @@
     <div
       :class="FSwitchClass"
       :style="{ background: modelValue ? openColor : closeColor }"
-      @click="changeSwitch"
+      @click="handleClick"
     >
       <span class="f-switch-roll" :style="rollStyle">
-        <f-icon v-if="icon" :icon="icon" />
+        <f-svg-icon v-if="icon" :icon="icon" :size="14" />
       </span>
     </div>